[dag] reorganize anchor election
